refactor: nav-items.ts 위치 변경
